remove putServiceRecipients from endpoints
