+++ app/pages/sections/service-recipients/controller.test.js
@@ -1,7 +1,7 @@
-import { getData, putData } from 'buying-catalogue-library';
-import { getServiceRecipientsContext, putServiceRecipients } from './controller';
+import { getData } from 'buying-catalogue-library';
+import { getServiceRecipientsContext } from './controller';
 import { logger } from '../../../logger';
-import { organisationApiUrl, orderApiUrl } from '../../../config';
+import { organisationApiUrl } from '../../../config';
 import * as contextCreator from './contextCreator';
 import { getRecipients as getRecipientsFromOrdapi } from '../../../helpers/api/ordapi/getRecipients';
 
@@ -25,19 +25,6 @@ const dataFromOrdapi = {
     name: 'Some service recipient 2',
     odsCode: 'ods2',
   }],
-};
-
-const mockData = {
-  _csrf: 'testCSRF',
-  XX1: 'Some service recipient 1',
-  XX2: 'Some service recipient 2',
-};
-
-const formattedMockData = {
-  serviceRecipients: [
-    { name: 'Some service recipient 1', odsCode: 'XX1' },
-    { name: 'Some service recipient 2', odsCode: 'XX2' },
-  ],
 };
 
 describe('service-recipients controller', () => {
@@ -112,38 +99,6 @@ describe('service-recipients controller', () => {
         serviceRecipientsData: dataFromOapi,
         selectedRecipientIdsData: dataFromOrdapi.serviceRecipients,
       });
-    });
-  });
-
-  describe('putServiceRecipients', () => {
-    beforeEach(() => {
-      putData.mockReset();
-    });
-
-    it('should call putData once with the correct params', async () => {
-      putData
-        .mockResolvedValueOnce({});
-
-      await putServiceRecipients({
-        orderId: 'order-id', data: mockData, accessToken: 'access_token',
-      });
-      expect(putData.mock.calls.length).toEqual(1);
-      expect(putData).toHaveBeenCalledWith({
-        endpoint: `${orderApiUrl}/api/v1/orders/order-id/sections/service-recipients`,
-        body: formattedMockData,
-        accessToken: 'access_token',
-        logger,
-      });
-    });
-
-    it('should return success as true if data is saved successfully', async () => {
-      putData
-        .mockResolvedValueOnce({});
-      const response = await putServiceRecipients({
-        orderId: 'order-id', data: mockData, accessToken: 'access_token',
-      });
-      expect(response.success).toEqual(true);
-      expect(response.errors).toEqual(undefined);
     });
   });
 });
